pyproject.toml
[tool.poetry]
name = "python-backend-template"
version = "1.4.2"
description = ""
authors = ["Artem Hrytsenko"]

[tool.poetry.dependencies]
python = "^3.11"

# FastAPI
fastapi = "^0.89.1"
pydantic = "^1.10.4"
python-multipart = "^0.0.5"
uvicorn = { extras = ["standard"], version = "^0.20.0" }

# Database
SQLAlchemy = {extras = ["mypy"], version = "^2.0.12"}
asyncpg = "^0.27.0"
alembic = "^1.9.2"
python-dateutil = "^2.8.2"

# S3
boto3 = "^1.26.128"

# Security
passlib = {extras = ['bcrypt'], version = '^1.7.4'}
python-jose = {extras = ['cryptography'], version = '^3.1.0'}
cryptography = "^3.3.1"
redis = {extras = ['hiredis'], version = "^4.5.4" }
pyfa-converter = "^1.0.3.0"

# Utils
python-slugify = "^8.0.1"
sentry-sdk = "^1.13.0"
pathos = "^0.3.0"
celery = "^5.2.7"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.2"
pycln = "^2.1.1"
flake8 = "^5.0.4"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^23.3.12"
flake8-builtins = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-eradicate = "^1.4.0"
flake8-pytest-style = "^1.6.0"
flake8-spellcheck = "^0.28.0"
pep8-naming = "^0.13.2"
SQLAlchemy = {extras = ["mypy"], version = "^2.0.12"}
pytest = "^7.2.1"
boto3-stubs = {extras = ['s3'], version = '^1.26.128'}
greenlet = "^2.0.2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.pycln]
expand-stars = true

[tool.mypy]
python_version = "3.11"
allow_redefinition = true
check_untyped_defs = true
disallow_any_explicit = false
disallow_subclassing_any = false
disallow_any_generics = true
disallow_untyped_calls = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = true
local_partial_types = true
strict_optional = true
strict_equality = true
show_error_codes = true
no_implicit_optional = true
warn_no_return = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = false
warn_return_any = true
strict = true
exclude = "venv|migrations"
plugins = "sqlalchemy.ext.mypy.plugin"

[tool.isort]
line_length = 80
atomic = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
skip_glob = "**/__init__.py"

[[tool.mypy.overrides]]
module = "src.lib.utils.redisgraph.*"
ignore_errors = true